Adds capability to export parameters from aws
* Adds lodash library
* Renames script from get_parameter_from_aws.js to
export_parameter_from_aws.js
