+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "aws-sdk": "^2.287.0",
+    "lodash": "^4.17.10",
     "yargs": "^12.0.1"
   },
   "name": "awsnodepostmanutils",
